third_party/blink/web_tests/external/wpt/css/css-inline/baseline-source/baseline-source-last-textarea-002.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] .target > * 1
  assert_equals: \n<span data-offset-x="45"></span>\noffsetLeft expected 45 but got 60
[FAIL] .target > * 3
  assert_equals: \n<span data-offset-x="45"></span>\noffsetLeft expected 45 but got 10
[FAIL] .target > * 4
  assert_equals: \n<textarea data-offset-x="10" class="inner">X X X X X</textarea>\noffsetLeft expected 10 but got 20
[FAIL] .target > * 5
  assert_equals: \n<span data-offset-x="45"></span>\noffsetLeft expected 45 but got 60
[FAIL] .target > * 7
  assert_equals: \n<span data-offset-x="45"></span>\noffsetLeft expected 45 but got 25
[FAIL] .target > * 9
  assert_equals: \n<span data-offset-x="45"></span>\noffsetLeft expected 45 but got 10
[FAIL] .target > * 10
  assert_equals: \n<textarea data-offset-x="10" class="inner" style="font-size: 100px;">X X X X X</textarea>\noffsetLeft expected 10 but got 20
[FAIL] .target > * 11
  assert_equals: \n<span data-offset-x="45"></span>\noffsetLeft expected 45 but got 10
Harness: the test ran to completion.
